refactor(57): move functions to files
add copyright header
